+++ integrations/streamer_test.go
@@ -9089,6 +9089,7 @@ stream
 			.group('{{ .ID }}')
 			.value('{{ index .Fields "count" }}')
 			.services('serviceA', 'serviceB', '{{ .Name }}')
+			.correlated('{{ .Name }}')
 `
 	tmInit := func(tm *kapacitor.TaskMaster) {
 		c := alerta.NewConfig()
@@ -9112,6 +9113,7 @@ stream
 				Text:        "kapacitor/cpu/serverA is CRITICAL @1971-01-01 00:00:10 +0000 UTC",
 				Origin:      "Kapacitor",
 				Service:     []string{"cpu"},
+				Correlate:   []string{"cpu"},
 				Timeout:     3600,
 			},
 		},
@@ -9126,6 +9128,7 @@ stream
 				Text:        "kapacitor/cpu/serverA is CRITICAL @1971-01-01 00:00:10 +0000 UTC",
 				Origin:      "override",
 				Service:     []string{"serviceA", "serviceB", "cpu"},
+				Correlate:   []string{"cpu"},
 				Value:       "10",
 				Timeout:     86400,
 			},